Renamed package to org.minecraftairshippirates.customwhitelist
